+++ packages/changesets/src/index.ts
@@ -1,0 +1,3 @@
+import { changelogFunctions } from "./changelog";
+
+export default changelogFunctions;