+++ packages/multistore/src/types.ts
@@ -1,7 +1,7 @@
 /**
  * Type alias for any store configuration.
  */
-export type StoreConfig = any;
+export type StoreConfig = Record<string, any>;
 
 /**
  * Cache manager is responsible for caching configuration and retreiving configuration from cache.
@@ -10,12 +10,12 @@ export interface CacheManager {
   /**
    * Gets a store configuration from cache storage based on the `key` value.
    */
-  get(key: string): StoreConfig;
+  get(key: string): StoreConfig | undefined;
 
   /**
    * Sets a store configuration in the cache storage with identifier equals `key` value.
    */
-  set(key: string, value: StoreConfig): StoreConfig;
+  set(key: string, value: StoreConfig): unknown;
 }
 
 /**
@@ -25,7 +25,7 @@ export interface MultistoreExtensionMethods {
   /**
    * Cache manager factory creates cache manager.
    */
-  cacheManagerFactory: (configuration: StoreConfig) => CacheManager;
+  cacheManagerFactory: () => CacheManager;
 
   /**
    * Overwrites base configuration with store configuration.
@@ -41,28 +41,6 @@ export interface MultistoreExtensionMethods {
   fetchConfiguration: (params: {
     domain: string;
   }) => Record<string, StoreConfig>;
-}
-
-export interface MiddlewareConfiguration {
-  /**
-   * Multistore configuration.
-   */
-  multistore: MultistoreExtensionMethods;
-
-  /**
-   * Other configration entries, that are not being used by the multistore extension.
-   */
-  [key: string]: any;
-}
-
-/**
- * Parameters containing multistore configuration used by the extension.
- */
-export interface ExtensionParams {
-  /**
-   * Middleware configuration.
-   */
-  configuration: MiddlewareConfiguration;
 }
 
 /**