+++ packages/sdk/src/__tests__/unit/buildModule.spec.ts
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import { extension1Mock } from "../__mocks__/extension1.mock";
 import module1Mock from "../__mocks__/module1.mock";
-import { buildModule } from "../../module/buildModule";
+import { buildModule } from "../../modules/buildModule";
 
 describe("[buildModule]", () => {
   it("should build module without extensions", () => {
@@ -26,7 +27,7 @@ describe("[buildModule]", () => {
       ...module1Mock({}),
       ...extension1Mock,
     };
-
+    const resolvedModule = module1Mock({});
     const result = buildModule(
       module1Mock,
       {},
@@ -35,8 +36,11 @@ describe("[buildModule]", () => {
     );
 
     expect(JSON.stringify(result)).toEqual(JSON.stringify(expected));
-    expect(extension1MockFn).toBeCalledTimes(1);
-    expect(extension1MockFn).toBeCalledWith(extensionOptions);
+    expect(extension1MockFn).toHaveBeenCalledTimes(1);
+    expect(extension1MockFn).toHaveBeenCalledWith(extensionOptions, {
+      methods: resolvedModule.connector,
+      context: resolvedModule.context,
+    });
   });
 
   it("module can be executed with options", () => {
@@ -46,7 +50,49 @@ describe("[buildModule]", () => {
       ...module1Mock({ test: true }),
       extend: {},
     };
-    expect(module1Mock).toBeCalledWith({ test: true });
+    expect(module1Mock).toHaveBeenCalledWith({ test: true });
     expect(JSON.stringify(result)).toEqual(JSON.stringify(expected));
+  });
+
+  it("should correctly infer types in typescript's strict mode", () => {
+    // No options
+    const moduleWithNoOptions = () => ({ connector: {} });
+    buildModule(moduleWithNoOptions);
+    buildModule(moduleWithNoOptions, {}, extension1Mock);
+    buildModule(moduleWithNoOptions, {}, () => extension1Mock, {
+      test: true,
+    });
+
+    // Optional options
+    const moduleWithOptionalOptions = (_opts?: { test: boolean }) => ({
+      connector: {},
+    });
+    buildModule(moduleWithOptionalOptions);
+    // @ts-expect-error - should correctly infer type of options
+    buildModule(moduleWithOptionalOptions, { test: "should be boolean" });
+    buildModule(moduleWithOptionalOptions, { test: true });
+    buildModule(moduleWithNoOptions, {}, extension1Mock);
+    buildModule(moduleWithNoOptions, {}, () => extension1Mock, {
+      test: true,
+    });
+
+    // Mandatory options
+    const moduleWithMandatoryOptions = (_opts: { test: boolean }) => ({
+      connector: {},
+    });
+    // @ts-expect-error - should require options as a second argument
+    buildModule(moduleWithMandatoryOptions);
+    // @ts-expect-error - should correctly infer type of options
+    buildModule(moduleWithOptionalOptions, { test: "should be boolean" });
+    buildModule(moduleWithMandatoryOptions, { test: true });
+    buildModule(moduleWithMandatoryOptions, { test: true }, extension1Mock);
+    buildModule(
+      moduleWithMandatoryOptions,
+      { test: true },
+      () => extension1Mock,
+      {
+        test: true,
+      }
+    );
   });
 });