+++ packages/sdk/src/modules/middlewareModule/consts.ts
@@ -1,0 +1,4 @@
+/**
+ * Symbol to mark a method config object.
+ */
+export const isConfig = Symbol("config");