+++ packages/sdk/src/modules/middlewareModule/index.ts
@@ -1,0 +1,3 @@
+export { middlewareModule } from "./module";
+export * from "./utils";
+export * from "./types";