+++ packages/storefront/shared/src/index.ts
@@ -1,3 +1,4 @@
-export * from "./__tests__/sdk";
+/* eslint-disable import/no-relative-packages */
+export * from "./__tests__/middleware/types";
 export * from "./helpers";
 export * from "./types";